+++ 6-calculator/src/utils/convertToPostfixNotation/convertToPostfixNotation.ts
